% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_cell_info.R
\docType{data}
\name{gene_cell_info}
\alias{gene_cell_info}
\title{Baseline expression information for genes across cell lines}
\format{A data frame with 1703457 observations on the following 6 variables.
\itemize{
\item{\code{pr_gene_symbol}}{a character vector}
\item{\code{cell}}{a character vector}
\item{\code{basex_affx}}{a numeric vector}
\item{\code{basex_rnaseq}}{a numeric vector}
\item{\code{copy_number}}{a numeric vector}
\item{\code{is_expressed}}{a logical vector}
}}
\usage{
data(gene_cell_info)
}
\description{
This data set contains baseline expression measurements for genes in cell lines [without any experimental perturbations] as profiles by LINCS
}
\keyword{datasets}

