% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KO_data.R
\docType{data}
\name{KO_data}
\alias{KO_data}
\title{LINCS knock-out data}
\format{A data frame with 36720 observations on the following 9 variables.
\itemize{
\item{\code{cell_id}}  {a character vector}
\item{\code{pert_desc}}  {a character vector}
\item{\code{pert_time}}  {a numeric vector}
\item{\code{dn100_bing}}  {a character vector}
\item{\code{dn100_full}}  {a character vector}
\item{\code{dn50_lm}}  {a character vector}
\item{\code{up100_bing}}  {a character vector}
\item{\code{up100_full}}  {a character vector}
\item{\code{up50_lm}}  {a character vector}
}}
\usage{
data(KO_data)
}
\value{
A data.frame object
}
\description{
This data set contains consensus genome signatures (CGS) 
that are the result of experimental knock-out studies conducted by the 
BROAD Institute.
}
\references{
http://lincsportal.ccs.miami.edu
}
\keyword{datasets}

