% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOdata.R
\docType{package}
\name{KOdata}
\alias{KOdata}
\alias{KOdata-package}
\title{KOdata: an R data package designed to be used with KEGGlincs}
\description{
KOdata: an R data package designed to be used with KEGGlincs
}

